fix(ProductCard): upselling kit renaming and export
