+++ packages/react/src/experimental/UpsellingKit/ProductCard/index.mdx
@@ -1,0 +1,17 @@
+import { Canvas, Meta, Unstyled, Controls } from "@storybook/blocks"
+import * as PresetStories from "./index.stories"
+
+<Meta of={PresetStories} />
+
+# Preset
+
+Customizable Product card to generate upsell request
+
+```tsx
+import { ProductCard } from "@factorialco/factorial-one/dist/experimental"
+```
+
+## Anatomy
+
+<Canvas of={PresetStories.Default} />
+<Controls of={PresetStories.Default} />
