+++ packages/react/src/experimental/UpsellingKit/ProductCard/index.tsx
@@ -1,7 +1,7 @@
 import { Button, IconType } from "@/factorial-one"
 import CrossIcon from "@/icons/app/Cross"
 import { useEffect, useState } from "react"
-import { ModuleAvatar } from "../Information/ModuleAvatar"
+import { ModuleAvatar } from "../../Information/ModuleAvatar"
 
 export type ProductCardProps = {
   title: string
@@ -45,7 +45,7 @@ export function ProductCard({
       <div>
         <div className="p-2">
           <div
-            className="flex h-[90px] w-auto cursor-pointer flex-row gap-2 rounded-md border-f1-border p-3 text-f1-foreground shadow-md hover:bg-f1-background-secondary"
+            className="flex h-auto w-auto cursor-pointer flex-row gap-2 rounded-md border-f1-border p-3 text-f1-foreground shadow-md hover:bg-f1-background-secondary"
             onClick={onClick}
           >
             <ModuleAvatar icon={icon} size="lg" />
