+++ packages/react/src/experimental/UpsellingKit/exports.tsx
@@ -1,0 +1,1 @@
+export * from "./ProductCard"
